+++ CHANGELOG.md
@@ -21,6 +21,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Introduced `AccessTokenInvalidException` for more specific error handling. **[BREAKING]**
 * Support for the `all` operator and passing arrays as `$value` in `Query::where()`.
 * Support for ordering by multiple fields.
+* The space metadata and the content types can now be cached with a CLI command.
+* Support for caching the Space and Content Types. The cache has to be manualy warmed and cleared.
 
 ### Changed
 * Changed the behavior of getting an array of links to not throw an exception when one of them has been deleted from the space. ([#19](https://github.com/contentful/contentful.php/pull/19))